api_watcher/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
deepdiff>=6.7.0
pyyaml>=6.0.1